+++ pkg/ingester/client/compat.go
@@ -328,6 +328,16 @@ func Fingerprint(labels labels.Labels) model.Fingerprint {
 	return model.Fingerprint(sum)
 }
 
+// LabelsToKeyString is used to form a string to be used as
+// the hashKey. Don't print, use l.String() for printing.
+func LabelsToKeyString(l labels.Labels) string {
+	// We are allocating 1024, even though most series are less than 600b long.
+	// But this is not an issue as this function is being inlined when called in a loop
+	// and buffer allocated is a static buffer and not a dynamic buffer on the heap.
+	b := make([]byte, 0, 1024)
+	return string(l.Bytes(b))
+}
+
 // MarshalJSON implements json.Marshaler.
 func (s Sample) MarshalJSON() ([]byte, error) {
 	t, err := json.Marshal(model.Time(s.TimestampMs))
