+++ pkg/ingester/client/compat_test.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"reflect"
 	"sort"
+	"strconv"
 	"testing"
 	"unsafe"
 
@@ -215,4 +216,54 @@ func verifyCollision(t *testing.T, collision bool, ls1 labels.Labels, ls2 labels
 	} else if !collision && Fingerprint(ls1) == Fingerprint(ls2) {
 		t.Errorf("expected different fingerprints for %v (%016x) and %v (%016x)", ls1.String(), Fingerprint(ls1), ls2.String(), Fingerprint(ls2))
 	}
+}
+
+// The main usecase for `LabelsToKeyString` is to generate hashKeys
+// for maps. We are benchmarking that here.
+func BenchmarkSeriesMap(b *testing.B) {
+	benchmarkSeriesMap(100000, b)
+}
+
+func benchmarkSeriesMap(numSeries int, b *testing.B) {
+	series := makeSeries(numSeries)
+	sm := make(map[string]int, numSeries)
+
+	b.ReportAllocs()
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		for i, s := range series {
+			sm[LabelsToKeyString(s)] = i
+		}
+
+		for _, s := range series {
+			_, ok := sm[LabelsToKeyString(s)]
+			if !ok {
+				b.Fatal("element missing")
+			}
+		}
+
+		if len(sm) != numSeries {
+			b.Fatal("the number of series expected:", numSeries, "got:", len(sm))
+		}
+	}
+}
+
+func makeSeries(n int) []labels.Labels {
+	series := make([]labels.Labels, 0, n)
+	for i := 0; i < n; i++ {
+		series = append(series, labels.FromMap(map[string]string{
+			"label0": "value0",
+			"label1": "value1",
+			"label2": "value2",
+			"label3": "value3",
+			"label4": "value4",
+			"label5": "value5",
+			"label6": "value6",
+			"label7": "value7",
+			"label8": "value8",
+			"label9": strconv.Itoa(i),
+		}))
+	}
+
+	return series
 }
